reshuffle VerticalCardListFeatureConnected files to match existing "ConnectedComponent" paradigm.
