+++ packages/apollos-ui-connected/src/VerticalCardListFeatureConnected/VerticalCardListFeature/VerticalCardListFeature.tests.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import { Providers } from '../utils/testUtils';
+import { Providers } from '../../utils/testUtils';
 
-import VerticalCardListFeature from './VerticalCardListFeature';
+import VerticalCardListFeature from '.';
 
 const cards = [
   {
